Re-themeing site
I wanted a bit more control over the overall layout of the page and the
colors, so I've overridden most of the theme from the minima theme and
replaced it with my own. The page is more colorful now. The experience
page looks almost identical to my resume/CV now too.